# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "microservice of commission-service",
        "title": "commission-service",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Check the health of the application",
                "responses": {
                    "200": {
                        "description": "Health check passed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/calculate": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a group of commission",
                "parameters": [
                    {
                        "description": "Array of uuid groups",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Product"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": "Syntax error in provided payload",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorAPI"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/calculate/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a group of commission",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The channel UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Product"
                        }
                    },
                    "400": {
                        "description": "Syntax error in provided payload",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorAPI"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorAPI": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "request.Product": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 3
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "response.Product": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}